requirements.txt
openai==0.28
chromadb
langchain
pypdf
tiktoken
ttkthemes
Pillow